Generalize partitioned writer (so we can use it for defillama)
