+++ src/op_analytics/coreutils/partitioned/writer.py
@@ -1,12 +1,16 @@
-from collections import defaultdict
 from dataclasses import dataclass
+from typing import Any
 
-from op_analytics.coreutils.logger import bound_contextvars, structlog, human_rows
+from overrides import override
 
+from op_analytics.coreutils.logger import bound_contextvars, structlog
+
+from .breakout import breakout_partitions
+from .dataaccess import init_data_access
 from .location import DataLocation
 from .output import ExpectedOutput, OutputData, OutputPartMeta
 from .status import all_outputs_complete
-from .writehelper import ParquetWriteManager
+from .writehelper import WriteManager
 
 log = structlog.get_logger()
 
@@ -17,6 +21,9 @@ class DataWriter:
 
     # Data store where we write the data.
     write_to: DataLocation
+
+    # Partition columns
+    partition_cols: list[str]
 
     # Markers Table
     markers_table: str
@@ -38,26 +45,8 @@ class DataWriter:
             markers_table=self.markers_table,
         )
 
-    def write_all(self, outputs: list[OutputData]):
-        """Write data and markers to all the specified locations.
-
-        The data is provided as a list of functions that return a dataframe. This lets us generalize
-        the way in which different tasks produce OutputDataFrame.
-        """
-
-        total_rows: dict[str, int] = defaultdict(int)
-
-        for output_data in outputs:
-            parts = self.write(self.write_to, output_data)
-
-            for part in parts:
-                total_rows[output_data.dataset_name] += part.row_count
-
-        summary = " ".join(f"{key}={human_rows(val)}" for key, val in total_rows.items())
-        summary = f"{self.write_to.name}::{summary}"
-        log.info(f"done writing. {summary}")
-
-    def write(self, location: DataLocation, output_data: OutputData) -> list[OutputPartMeta]:
+    def write(self, output_data: OutputData) -> list[OutputPartMeta]:
+        """Write data and corresponding marker."""
         expected_output = self.expected_outputs[output_data.dataset_name]
 
         # The default partition value is included in logs because it includes
@@ -67,11 +56,42 @@ class DataWriter:
         with bound_contextvars(
             dataset=output_data.dataset_name, **(output_data.default_partition or {})
         ):
-            manager = ParquetWriteManager(
-                location=location,
+            manager = PartitionedWriteManager(
+                partition_cols=self.partition_cols,
+                location=self.write_to,
                 expected_output=expected_output,
                 markers_table=self.markers_table,
                 force=self.force,
             )
 
             return manager.write(output_data)
+
+
+@dataclass
+class PartitionedWriteManager(WriteManager):
+    partition_cols: list[str]
+
+    @override
+    def write_implementation(self, output_data: Any) -> list[OutputPartMeta]:
+        assert isinstance(output_data, OutputData)
+
+        client = init_data_access()
+
+        parts = breakout_partitions(
+            df=output_data.dataframe,
+            partition_cols=self.partition_cols,
+            default_partition=output_data.default_partition,
+        )
+
+        parts_meta: list[OutputPartMeta] = []
+        for part in parts:
+            client.write_single_part(
+                location=self.location,
+                dataframe=part.df,
+                full_path=part.meta.full_path(
+                    self.expected_output.root_path, self.expected_output.file_name
+                ),
+            )
+            parts_meta.append(part.meta)
+
+        return parts_meta
