+++ src/op_analytics/datapipeline/etl/intermediate/main.py
@@ -92,10 +92,7 @@ def executor(task: IntermediateModelsTask) -> None:
                     )
 
                 for result_name, rel in model_results.items():
-                    location = task.data_writer.write_to
-                    log.info("writing model", result=result_name, location=location)
                     task.data_writer.write(
-                        location=location,
                         output_data=OutputData(
                             dataframe=rel.pl(),
                             dataset_name=f"{model_name}/{result_name}",
